fix(shields): drop errexit toggles in E2E, move drift tests to focused file
