+++ src/lib/shields/index.test.ts
@@ -656,115 +656,6 @@ describe("shields — unit logic", () => {
     });
   });
 
-  // -------------------------------------------------------------------
-  // shieldsStatus: locked-state drift surface
-  // -------------------------------------------------------------------
-  describe("shieldsStatus surfaces drift returned by the verifier", () => {
-    function writeLockedState(sandboxName: string): void {
-      fs.mkdirSync(stateDir(), { recursive: true });
-      fs.writeFileSync(
-        path.join(stateDir(), `shields-${sandboxName}.json`),
-        JSON.stringify(
-          {
-            shieldsDown: false,
-            updatedAt: new Date().toISOString(),
-          },
-          null,
-          2,
-        ),
-        { mode: 0o600 },
-      );
-    }
-
-    it("prints DRIFTED with the issue list and exits 2 when the verifier reports drift", async () => {
-      const sandboxName = "openclaw";
-      writeLockedState(sandboxName);
-      const driftIssues = [
-        "/sandbox/.openclaw/openclaw.json mode=660 (expected 444)",
-        "/sandbox/.openclaw/openclaw.json owner=sandbox:sandbox (expected root:root)",
-        "dir mode=2770 (expected 755)",
-        "dir owner=sandbox:sandbox (expected root:root)",
-      ];
-      const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-      const exitSpy = vi
-        .spyOn(process, "exit")
-        .mockImplementation((code?: string | number | null) => {
-          throw new Error(`exit ${String(code)}`);
-        });
-
-      const { shieldsStatus } = await loadShieldsModule();
-      expect(() =>
-        shieldsStatus(sandboxName, true, {
-          verifyLockState: () => ({ ok: false, issues: driftIssues }),
-          resolveConfig: () => ({
-            agentName: "openclaw",
-            configPath: "/sandbox/.openclaw/openclaw.json",
-            configDir: "/sandbox/.openclaw",
-          }),
-        }),
-      ).toThrow("exit 2");
-
-      expect(errorSpy).toHaveBeenCalledWith(
-        "  Shields: UP (DRIFTED — declared locked but sandbox filesystem differs)",
-      );
-      expect(errorSpy).toHaveBeenCalledWith("  Drift:");
-      for (const issue of driftIssues) {
-        expect(errorSpy).toHaveBeenCalledWith(`    - ${issue}`);
-      }
-      expect(errorSpy).toHaveBeenCalledWith(
-        `  Recovery: nemoclaw ${sandboxName} shields up   # re-lock and re-verify`,
-      );
-      expect(exitSpy).toHaveBeenCalledWith(2);
-    });
-
-    it("prints a clean locked status when the verifier reports no drift", async () => {
-      const sandboxName = "openclaw";
-      writeLockedState(sandboxName);
-      const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
-      const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-
-      const { shieldsStatus } = await loadShieldsModule();
-      shieldsStatus(sandboxName, true, {
-        verifyLockState: () => ({ ok: true, issues: [] }),
-        resolveConfig: () => ({
-          agentName: "openclaw",
-          configPath: "/sandbox/.openclaw/openclaw.json",
-          configDir: "/sandbox/.openclaw",
-        }),
-      });
-
-      expect(logSpy).toHaveBeenCalledWith("  Shields: UP (lockdown active)");
-      expect(logSpy).toHaveBeenCalledWith("  Policy:  restrictive");
-      expect(errorSpy).not.toHaveBeenCalled();
-    });
-
-    it("treats a resolveConfig throw as drift so the locked status cannot mask a setup gap", async () => {
-      const sandboxName = "openclaw";
-      writeLockedState(sandboxName);
-      const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-      const exitSpy = vi
-        .spyOn(process, "exit")
-        .mockImplementation((code?: string | number | null) => {
-          throw new Error(`exit ${String(code)}`);
-        });
-
-      const { shieldsStatus } = await loadShieldsModule();
-      expect(() =>
-        shieldsStatus(sandboxName, true, {
-          verifyLockState: () => ({ ok: true, issues: [] }),
-          resolveConfig: () => {
-            throw new Error("agent config not found");
-          },
-        }),
-      ).toThrow("exit 2");
-
-      const allErrors = errorSpy.mock.calls.map((args) => args[0]).join("\n");
-      expect(allErrors).toContain(
-        "unable to resolve agent config target: agent config not found",
-      );
-      expect(exitSpy).toHaveBeenCalledWith(2);
-    });
-  });
 });
 
 // -------------------------------------------------------------------
